PivotGrid: A little refactoring of eslintrc + module names.
