+++ js/__internal/grids/pivot_grid/chart_integration/module.ts
@@ -2,7 +2,7 @@ import $ from '@js/core/renderer';
 import { extend } from '@js/core/utils/extend';
 import { each } from '@js/core/utils/iterator';
 
-import { foreachTree, formatValue, createPath } from '../widget_utils';
+import { foreachTree, formatValue, createPath } from '../module_widget_utils';
 
 const FORMAT_DICTIONARY = {
   number: 'numeric',
