+++ js/ui/pivot_grid/data_source.utils.js
@@ -1,2 +1,2 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/data_source/module_utils';
