+++ js/ui/pivot_grid/local_store.js
@@ -1,4 +1,4 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/local_store/module';
