+++ js/ui/pivot_grid/remote_store.js
@@ -1,2 +1,2 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/remote_store/module';
