+++ js/ui/pivot_grid/remote_store.utils.js
@@ -1,4 +1,4 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/remote_store/module_utils';
