+++ js/ui/pivot_grid/ui.pivot_grid.area_item.js
@@ -1,3 +1,3 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/area_item/module';
