+++ js/ui/pivot_grid/ui.pivot_grid.chart_integration.js
@@ -1,3 +1,3 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/chart_integration/module';
