+++ js/ui/pivot_grid/ui.pivot_grid.data_controller.js
@@ -1,4 +1,4 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/data_controller/module';
