+++ js/ui/pivot_grid/ui.pivot_grid.field_chooser_base.js
@@ -1,4 +1,4 @@
 import { FieldChooserBase } from '../../__internal/grids/pivot_grid/field_chooser/module_base';
-/** @deprecated **/
+// deprecated
 export default FieldChooserBase;
