+++ js/ui/pivot_grid/ui.pivot_grid.fields_area.js
@@ -1,3 +1,3 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/fields_area/module';
