+++ js/ui/pivot_grid/ui.pivot_grid.headers_area.js
@@ -1,4 +1,4 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/headers_area/module';
