+++ js/ui/pivot_grid/ui.pivot_grid.js
@@ -1,4 +1,4 @@
-import { PivotGrid } from '../../__internal/grids/pivot_grid/widget';
+import { PivotGrid } from '../../__internal/grids/pivot_grid/module_widget';
 export default PivotGrid;
