+++ js/ui/pivot_grid/ui.pivot_grid.summary_display_modes.js
@@ -1,3 +1,3 @@
-/** @deprecated **/
+// deprecated
 export * from '../../__internal/grids/pivot_grid/summary_display_modes/module';
