+++ js/ui/pivot_grid/ui.pivot_grid.utils.js
@@ -1,3 +1,3 @@
-/** @deprecated **/
-export * from '../../__internal/grids/pivot_grid/widget_utils';
+// deprecated
+export * from '../../__internal/grids/pivot_grid/module_widget_utils';
