+++ js/ui/pivot_grid/ui.sortable.js
@@ -1,4 +1,4 @@
 import { Sortable } from '../../__internal/grids/pivot_grid/sortable/module';
-/** @deprecated **/
+// deprecated
 export default Sortable;
