+++ testing/tests/DevExpress.ui.widgets.pivotGrid/dataController.tests.js
@@ -3,7 +3,7 @@ import Class from 'core/class';
 import { DataController } from '__internal/grids/pivot_grid/data_controller/module';
 import virtualScrolling from 'ui/grid_core/ui.grid_core.virtual_scrolling_core';
 import stateStoring from 'ui/grid_core/ui.grid_core.state_storing_core';
-import pivotGridUtils from '__internal/grids/pivot_grid/widget_utils';
+import pivotGridUtils from '__internal/grids/pivot_grid/module_widget_utils';
 import { PivotGridDataSource } from '__internal/grids/pivot_grid/data_source/module';
 
 import executeAsyncMock from '../../helpers/executeAsyncMock.js';
