+++ testing/tests/DevExpress.ui.widgets.pivotGrid/pivotGrid.markup.tests.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import windowUtils from 'core/utils/window';
 
-import '__internal/grids/pivot_grid/widget';
+import '__internal/grids/pivot_grid/module_widget';
 
 QUnit.module('PivotGrid markup tests', () => {
