+++ testing/tests/DevExpress.ui.widgets.pivotGrid/store.remote.tests.js
@@ -2,7 +2,7 @@ import $ from 'jquery';
 import { RemoteStore } from '__internal/grids/pivot_grid/remote_store/module';
 import { forEachGroup } from '__internal/grids/pivot_grid/remote_store/module_utils';
 import { sort } from '__internal/grids/pivot_grid/data_source/module_utils';
-import pivotGridUtils from '__internal/grids/pivot_grid/widget_utils';
+import pivotGridUtils from '__internal/grids/pivot_grid/module_widget_utils';
 import ArrayStore from 'data/array_store';
 import DataSource from 'data/data_source';
